pyproject.toml
[project]
name = "mlflow-test"
version = "0.1.0"
description = ""
authors = [
    {name = "genii01",email = "genii0627@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "onnx (>=1.17.0,<2.0.0)",
    "onnxruntime (>=1.20.1,<2.0.0)",
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psutil (>=6.1.1,<7.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.10.5,<3.0.0)",
    "datetime (>=5.5,<6.0)",
    "mlflow (>=2.19.0,<3.0.0)",
    "boto3 (>=1.36.3,<2.0.0)",
    "botocore (>=1.36.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
